chore(functional-tests): simplify and speed up execution of functional tests Because:
* The average 90-d functional test job time exceeded 10 minutes on stage

This commit:

* Simplifies all functional test runs to use an xlarge instances with parallelism set to 8
* Splits test modules to allow circleci to distribute tests more evenly accross runners

Closes #FXA-10941
