+++ packages/functional-tests/tests/misc/recoveryKeyPromoInline.spec.ts
@@ -6,56 +6,6 @@ import { expect, test } from '../../lib/fixtures/standard';
 import { getCode } from 'packages/fxa-settings/src/lib/totp';
 
 test.describe('recovery key promo', () => {
-  test.describe('settings banner', () => {
-    test('can setup recovery key from settings promo banner', async ({
-      target,
-      pages: { page, inlineRecoveryKey, signin, settings, recoveryKey },
-      testAccountTracker,
-    }) => {
-      const credentials = await testAccountTracker.signUp();
-      await page.goto(target.contentServerUrl);
-      await signin.fillOutEmailFirstForm(credentials.email);
-      await signin.fillOutPasswordForm(credentials.password);
-
-      await expect(settings.settingsHeading).toBeVisible();
-      await expect(settings.recoveryKey.status).toHaveText('Not Set');
-
-      await inlineRecoveryKey.getBannerCreateLink().click();
-
-      await recoveryKey.acknowledgeInfoForm();
-      await recoveryKey.fillOutConfirmPasswordForm(credentials.password);
-
-      await expect(recoveryKey.recoveryKeyCreatedHeading).toBeVisible();
-
-      // Notification banner is no longer visible
-      await expect(inlineRecoveryKey.getBannerCreateLink()).not.toBeVisible();
-    });
-
-    test('can dismiss', async ({
-      target,
-      pages: { page, inlineRecoveryKey, signin, settings, recoveryKey },
-      testAccountTracker,
-    }) => {
-      const credentials = await testAccountTracker.signUp();
-      await page.goto(target.contentServerUrl);
-      await signin.fillOutEmailFirstForm(credentials.email);
-      await signin.fillOutPasswordForm(credentials.password);
-
-      await expect(settings.settingsHeading).toBeVisible();
-      await expect(settings.recoveryKey.status).toHaveText('Not Set');
-
-      await inlineRecoveryKey.getBannerCreateLink().click();
-
-      await recoveryKey.acknowledgeInfoForm();
-      await recoveryKey.fillOutConfirmPasswordForm(credentials.password);
-
-      await expect(recoveryKey.recoveryKeyCreatedHeading).toBeVisible();
-
-      // Notification banner is no longer visible
-      await expect(inlineRecoveryKey.getBannerCreateLink()).not.toBeVisible();
-    });
-  });
-
   test.describe('inline', () => {
     test.beforeEach(async ({ pages: { configPage } }, { project }) => {
       const config = await configPage.getConfig();
