+++ packages/functional-tests/tests/settings/totp.spec.ts
@@ -9,7 +9,6 @@ import { TestAccountTracker } from '../../lib/testAccountTracker';
 import { SettingsPage } from '../../pages/settings';
 import { TotpCredentials, TotpPage } from '../../pages/settings/totp';
 import { SigninPage } from '../../pages/signin';
-import { SigninRecoveryCodePage } from '../../pages/signinRecoveryCode';
 import { SigninTotpCodePage } from '../../pages/signinTotpCode';
 
 test.describe('severity-1 #smoke', () => {
@@ -91,192 +90,6 @@ test.describe('severity-1 #smoke', () => {
       await expect(settings.totp.status).toHaveText('Enabled');
 
       await settings.disconnectTotp(); // Required before teardown
-    });
-
-    test('totp valid recovery code', async ({
-      target,
-      pages: {
-        page,
-        settings,
-        signin,
-        signinRecoveryCode,
-        signinTotpCode,
-        totp,
-      },
-      testAccountTracker,
-    }) => {
-      const credentials = await signInAccount(
-        target,
-        page,
-        settings,
-        signin,
-        testAccountTracker
-      );
-
-      await settings.goto();
-      const { recoveryCodes } = await addTotp(settings, totp);
-      await settings.signOut();
-      await signinWithRecoveryCode(
-        credentials,
-        recoveryCodes[0],
-        page,
-        settings,
-        signin,
-        signinRecoveryCode,
-        signinTotpCode
-      );
-      await expect(settings.settingsHeading).toBeVisible();
-      await settings.disconnectTotp();
-    });
-
-    test('totp invalid recovery code', async ({
-      target,
-      pages: {
-        page,
-        settings,
-        signin,
-        signinRecoveryCode,
-        signinTotpCode,
-        totp,
-      },
-      testAccountTracker,
-    }) => {
-      const credentials = await signInAccount(
-        target,
-        page,
-        settings,
-        signin,
-        testAccountTracker
-      );
-      await settings.goto();
-      const { recoveryCodes } = await addTotp(settings, totp);
-      await settings.signOut();
-      await signin.fillOutEmailFirstForm(credentials.email);
-      await signin.fillOutPasswordForm(credentials.password);
-      await expect(page).toHaveURL(/signin_totp_code/);
-      await signinTotpCode.useRecoveryCodeLink.click();
-      await expect(page).toHaveURL(/signin_recovery_code/);
-      await signinRecoveryCode.fillOutCodeForm('invalid!!!');
-      await expect(
-        page.getByText('Invalid backup authentication code')
-      ).toBeVisible();
-
-      // Required before teardown
-      await signinRecoveryCode.fillOutCodeForm(recoveryCodes[0]);
-      await expect(settings.settingsHeading).toBeVisible();
-      await settings.disconnectTotp();
-    });
-
-    test('can change backup authentication codes', async ({
-      target,
-      pages: {
-        page,
-        settings,
-        signin,
-        signinTotpCode,
-        signinRecoveryCode,
-        totp,
-      },
-      testAccountTracker,
-    }) => {
-      const credentials = await signInAccount(
-        target,
-        page,
-        settings,
-        signin,
-        testAccountTracker
-      );
-
-      await settings.goto();
-      const { recoveryCodes } = await addTotp(settings, totp);
-      await settings.totp.changeButton.click();
-
-      const newCodes = await totp.getRecoveryCodes();
-      expect(newCodes.some((c) => recoveryCodes.includes(c))).toBe(false);
-
-      await totp.step2ContinueButton.click();
-      await totp.step3RecoveryCodeTextbox.fill(newCodes[0]);
-      await totp.step3FinishButton.click();
-
-      await expect(settings.settingsHeading).toBeVisible();
-      await expect(settings.alertBar).toHaveText(
-        'Account backup authentication codes updated'
-      );
-
-      await settings.signOut();
-      await signinWithRecoveryCode(
-        credentials,
-        newCodes[0],
-        page,
-        settings,
-        signin,
-        signinRecoveryCode,
-        signinTotpCode
-      );
-
-      await expect(settings.settingsHeading).toBeVisible();
-      await settings.disconnectTotp(); // Required before teardown
-    });
-
-    test('can get new backup authentication codes via email', async ({
-      target,
-      pages: {
-        page,
-        settings,
-        signin,
-        signinTotpCode,
-        signinRecoveryCode,
-        totp,
-      },
-      testAccountTracker,
-    }) => {
-      const credentials = await signInAccount(
-        target,
-        page,
-        settings,
-        signin,
-        testAccountTracker
-      );
-
-      await settings.goto();
-      const { recoveryCodes } = await addTotp(settings, totp);
-      await settings.signOut();
-
-      for (let i = 0; i < recoveryCodes.length - 3; i++) {
-        await signinWithRecoveryCode(
-          credentials,
-          recoveryCodes[i],
-          page,
-          settings,
-          signin,
-          signinRecoveryCode,
-          signinTotpCode
-        );
-        await expect(settings.settingsHeading).toBeVisible();
-        await settings.signOut();
-      }
-
-      await signinWithRecoveryCode(
-        credentials,
-        recoveryCodes[recoveryCodes.length - 1],
-        page,
-        settings,
-        signin,
-        signinRecoveryCode,
-        signinTotpCode
-      );
-
-      await expect(settings.settingsHeading).toBeVisible();
-      const link = await target.emailClient.getLowRecoveryLink(
-        credentials.email
-      );
-      await page.goto(link);
-      const newCodes = await totp.getRecoveryCodes();
-      expect(newCodes.length).toEqual(recoveryCodes.length);
-
-      // Disconnect totp, required before teardown
-      await totp.step1CancelButton.click();
-      await settings.disconnectTotp();
     });
 
     test('delete account with totp enabled', async ({
@@ -365,21 +178,4 @@ async function signinWithTotp(
   const code = await getCode(secret);
   await signinTotpCode.fillOutCodeForm(code);
   await expect(settings.settingsHeading).toBeVisible();
-}
-
-async function signinWithRecoveryCode(
-  credentials: Credentials,
-  recoveryCode: string,
-  page: Page,
-  settings: SettingsPage,
-  signin: SigninPage,
-  signinRecoveryCode: SigninRecoveryCodePage,
-  signinTotpCode: SigninTotpCodePage
-): Promise<void> {
-  await signin.fillOutEmailFirstForm(credentials.email);
-  await signin.fillOutPasswordForm(credentials.password);
-  await expect(page).toHaveURL(/signin_totp_code/);
-  await signinTotpCode.useRecoveryCodeLink.click();
-  await expect(page).toHaveURL(/signin_recovery_code/);
-  await signinRecoveryCode.fillOutCodeForm(recoveryCode);
 }
